requirements.optimized.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
# torch will be installed separately with CPU-only version
torchvision>=0.22.1
ultralytics>=8.0.0
transformers>=4.30.0
httpx>=0.25.0
aiohttp>=3.9.0
gunicorn>=21.2.0
